refactor(Display): 값을 입/출력 형태로 변환한다
